extract table manager trait
