+++ src/moonlink/src/storage/iceberg.rs
@@ -9,6 +9,7 @@ pub(super) mod parquet_stats_utils;
 pub(super) mod parquet_utils;
 pub(super) mod puffin_utils;
 pub(super) mod puffin_writer_proxy;
+pub(super) mod table_manager;
 pub(super) mod table_property;
 pub(super) mod tokio_retry_utils;
 pub(super) mod utils;
