+++ src/moonlink/src/storage/iceberg/state_tests.rs
@@ -31,7 +31,7 @@ use iceberg::Result as IcebergResult;
 use crate::row::MoonlinkRow;
 use crate::row::RowValue;
 use crate::storage::iceberg::iceberg_table_manager::IcebergTableManager;
-use crate::storage::iceberg::iceberg_table_manager::TableManager;
+use crate::storage::iceberg::table_manager::TableManager;
 use crate::storage::iceberg::test_utils::{
     check_deletion_vector_consistency_for_snapshot, create_table_and_iceberg_manager,
     load_arrow_batch, validate_recovered_snapshot,
